Fix unsigned integer error

+++ Practice.cpp
@@ -32,7 +32,7 @@ void Practice::sortDescending(int & first, int & second, int & third)
 // character in the string, but disregarding case ('x' is the same as 'X')
 bool Practice::isPalindrome(string input)
 {
-  for(int i=0; i < input.size(); i++)
+  for(int i=0; i < (int)input.size(); i++)
   {
     if( input[i] < 'A' || input[i] > 'Z' )
     {
@@ -40,7 +40,7 @@ bool Practice::isPalindrome(string input)
       input[i] = input[i] - ('a' - 'A');
     }
   }
-  for(int i=0; i < input.size()/2; i++)
+  for(int i=0; i < (int)input.size()/2; i++)
   {
     if( input[i] != input[input.size()-1-i] )
       return false;
